requirements_ultimate.txt
# VUC-2026 Ultimate System Requirements
# Complete dependency list for optimized AI-powered YouTube Empire

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Google AI & OAuth
google-generativeai==0.3.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
googleapiclient==2.108.0

# ElevenLabs Voice API
elevenlabs==0.2.26

# Database & Storage
asyncpg==0.29.0
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.1

# Task Queue
celery==5.3.4
flower==2.0.1

# HTTP & Networking
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Media Processing
ffmpeg-python==0.2.0
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3

# Security & Authentication
PyJWT==2.8.0
bcrypt==4.1.2
cryptography==41.0.8
passlib[bcrypt]==1.7.4

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Monitoring & Analytics
psutil==5.9.6
GPUtil==1.4.0
prometheus-client==0.19.0

# AI & ML
openai==1.3.7
anthropic==0.7.8
transformers==4.36.0
torch==2.1.1

# Data Processing
pandas==2.1.4
numpy==1.24.3

# File Handling
aiofiles==23.2.1
python-magic==0.4.27

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Logging
structlog==23.2.0
colorlog==6.8.0

# Time & Dates
python-dateutil==2.8.2
pytz==2023.3

# JSON & Serialization
orjson==3.9.10
ujson==5.8.0

# Caching
aiocache==0.12.2
cachetools==5.3.2

# Rate Limiting
slowapi==0.1.9
limits==3.7.0

# Validation
email-validator==2.1.0
python-email-validator==1.3.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# WebSocket
websockets==12.0

# Background Tasks
apscheduler==3.10.4

# File Storage
boto3==1.34.0
google-cloud-storage==2.10.0

# Image Processing
cairosvg==2.7.1
wand==0.6.13

# Audio Processing
pydub==0.25.1
librosa==0.10.1

# Video Processing
moviepy==1.0.3
imageio==2.31.6

# Text Processing
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1

# Web Scraping
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0

# Database Migrations
alembic==1.13.1

# API Documentation
redoc==2.0.0
swagger-ui-bundle==0.0.9

# Performance
uvloop==0.19.0
orjson==3.9.10

# Utilities
tqdm==4.66.1
click==8.1.7
rich==13.7.0
typer==0.9.0

# Memory Optimization
memory-profiler==0.61.0
pympler==0.9

# Async Utilities
async-timeout==4.0.3
asyncstdlib==3.10.9

# Error Handling
sentry-sdk==1.38.0

# Configuration Management
dynaconf==3.2.4

# Template Engine
jinja2==3.1.2

# Email
aiosmtplib==3.0.1

# GraphQL
strawberry-graphql==0.215.1

# gRPC
grpcio==1.59.3
grpcio-tools==1.59.3

# Message Queue
kombu==5.3.4
pika==1.3.2

# Search
elasticsearch==8.11.0
whoosh==2.7.4

# Cache Backend
django-redis==5.4.0
hiredis==2.2.3

# Task Scheduling
celery-beat==2.5.0
django-celery-beat==2.5.0

# API Rate Limiting
django-ratelimit==4.1.0
ratelimit==2.2.1

# Data Validation
marshmallow==3.20.1
cerberus==1.3.5

# HTTP Client
httpcore==0.18.0
httpx==0.25.2

# WebSocket Support
websockets==12.0
socketio==5.10.0

# Real-time
socket.io-client==0.6.1

# File Upload
python-multipart==0.0.6
form-data==1.0.1

# Compression
gzip==1.0
lz4==4.3.2
zstandard==0.22.0

# Encryption
cryptography==41.0.8
pycryptodome==3.19.0

# Hashing
hashlib-compat==1.0.1
passlib==1.7.4

# Token Management
pyjwt==2.8.0
jose==1.0.0

# Session Management
itsdangerous==2.1.2
flask-session==0.5.0

# CORS
fastapi-cors==0.0.6
flask-cors==4.0.0

# Static Files
aiofiles==23.2.1
python-static-files==1.0.0

# Template Rendering
jinja2==3.1.2
mako==1.3.0

# Markdown Processing
markdown==3.5.1
markdownify==0.12.1

# URL Routing
pathlib2==2.3.7
furl==2.1.3

# Environment Variables
environs==10.0.0
configparser==6.0.0

# Logging Enhancement
loguru==0.7.2
structlog==23.2.0

# Performance Monitoring
py-spy==0.3.14
memory-profiler==0.61.0

# Database Connection Pool
sqlalchemy==2.0.23
asyncpg==0.29.0

# Redis Client
redis==5.0.1
aioredis==2.0.1

# Message Broker
celery==5.3.4
kombu==5.3.4

# Task Management
dramatiq==1.15.0
rq==1.15.1

# API Gateway
fastapi-gateway==0.0.1
api-gateway==1.0.0

# Load Balancing
nginx==1.24.0
haproxy==2.8.0

# Container Support
docker==6.1.3
docker-compose==1.29.2

# Cloud Services
boto3==1.34.0
google-cloud==0.34.0
azure-storage-blob==12.19.0

# Development Tools
pre-commit==3.6.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation Tools
sphinx==7.2.6
mkdocs==1.5.3
sphinx-rtd-theme==2.0.0

# Testing Framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Performance Testing
locust==2.17.0
wrk==4.2.0

# Security Tools
bandit==1.7.5
safety==2.3.5
vulture==2.10

# Code Quality
pylint==3.0.3
pycodestyle==2.11.1
autopep8==2.0.4

# IDE Support
jupyter==1.0.0
ipython==8.17.2

# Package Management
pip==23.3.2
setuptools==69.0.3
wheel==0.42.0
